+++ data/crac/crac-util/src/test/java/com/farao_community/farao/data/crac_util/CracCleanerTest.java
@@ -5,22 +5,21 @@
  *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.farao_community.farao.data.crac_io_json;
+package com.farao_community.farao.data.crac_util;
 
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_io_api.CracImporters;
 import com.powsybl.iidm.import_.Importers;
 import com.powsybl.iidm.network.Network;
-
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
 /**
- *
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  */
-public class CracAliasesUtilTest {
+
+public class CracCleanerTest {
 
     @Test
     public void testCracAliasesUtil() {
@@ -30,7 +29,7 @@ public class CracAliasesUtilTest {
         network.getBranch("FFR2AA1H DDE3AA1F 1").addAlias("FFR2AA1H DDE3AA1F HFSK JDV");
         network.getBranch("DDE1AA1D DDE2AA1E 1").addAlias("DDE2AA1E DDE1AA1D DLJKSC H");
 
-        CracImporters.cracAliasesUtil(crac, network);
+        CracCleaner.cracAliasesUtil(crac, network);
 
         assertEquals(1, network.getBranch("FFR1AA1G FFR3AA1I 1").getAliases().size());
         assertEquals(2, network.getBranch("DDE1AA1D DDE2AA1E 1").getAliases().size());
