+++ data/crac/crac-util/src/test/resources/case-for-aliases.uct
