+++ data/crac/crac-util/src/test/resources/crac-for-aliases.json
