+++ flowbased-computation/flowbased-computation-api/pom.xml
@@ -29,6 +29,11 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>farao-crac-util</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>farao-flowbased-domain</artifactId>
             <version>${project.version}</version>
         </dependency>
